feat: move builder cli to lib.rs - adds tests - closes #39
